+++ widgets/toolbars/options/kpToolWidgetBase_ribbon.cpp
@@ -9,52 +9,55 @@
 
 #include <QItemSelectionModel>
 
-kpToolWidgetBase_Ribbon::kpToolWidgetBase_Ribbon (QWidget *parent, const QString &name) :
-    kpToolWidgetBase(parent)
+kpToolWidgetBase::kpToolWidgetBase (QWidget *parent, const QString &name) :
+    SARibbonGallery(parent)
 {
-    m_gallery = new SARibbonGallery(this);
-    m_gallery->setSizePolicy (QSizePolicy::Maximum, QSizePolicy::Maximum);
-    m_galleryGroup = m_gallery->addGalleryGroup();
+    // m_gallery->setSizePolicy (QSizePolicy::Maximum, QSizePolicy::Maximum);
+    m_galleryGroup = SARibbonGallery::addGalleryGroup();
+    m_galleryGroup->setGroupTitle(name);
+    // m_galleryGroup->setDisplayRow(SARibbonGalleryGroup::DisplayRow::DisplayTwoRow);
 
     m_galleryGroup->setupGroupModel();
 
-    connect(m_galleryGroup->selectionModel(), &QItemSelectionModel::selectionChanged, this, &kpToolWidgetBase_Ribbon::callOptionSelected);
+    connect(this, &SARibbonGallery::triggered, this, &kpToolWidgetBase::callOptionSelected);
+    // connect(m_galleryGroup->selectionModel(), &QItemSelectionModel::selectionChanged, this, &kpToolWidgetBase::callOptionSelected);
 
     setObjectName (name);
 }
 
-kpToolWidgetBase_Ribbon::~kpToolWidgetBase_Ribbon () = default;
+kpToolWidgetBase::~kpToolWidgetBase () = default;
 
-void kpToolWidgetBase_Ribbon::addOption (const QPixmap &pixmap, const QString &toolTip)
+void kpToolWidgetBase::addOption (const QPixmap &pixmap, const QString &toolTip)
 {
     m_galleryGroup->addItem(toolTip, pixmap);
+    m_galleryGroup->recalcGridSize();
 }
 
-void kpToolWidgetBase_Ribbon::startNewOptionRow ()
+void kpToolWidgetBase::startNewOptionRow ()
 {
     Q_ASSERT(false);    // Code that hasn't been refactored away from using this function will crash. Even if it thinks it's using the old brush selection widget, it must treat it as a 1d widget and not a 2d widget.
 }
 
-void kpToolWidgetBase_Ribbon::finishConstruction (int fallBackRow, int fallBackCol)
+void kpToolWidgetBase::finishConstruction (int fallBackRow, int fallBackCol)
 {
 }
 
-QPair <int, int> kpToolWidgetBase_Ribbon::defaultSelectedRowAndCol () const
+QPair <int, int> kpToolWidgetBase::defaultSelectedRowAndCol () const
 {
     return qMakePair(0, 0);
 }
 
-int kpToolWidgetBase_Ribbon::defaultSelectedRow () const
+int kpToolWidgetBase::defaultSelectedRow () const
 {
     return defaultSelectedRowAndCol ().first;
 }
 
-int kpToolWidgetBase_Ribbon::defaultSelectedCol () const
+int kpToolWidgetBase::defaultSelectedCol () const
 {
     return defaultSelectedRowAndCol ().second;
 }
 
-void kpToolWidgetBase_Ribbon::saveSelectedAsDefault () const
+void kpToolWidgetBase::saveSelectedAsDefault () const
 {
     if (objectName ().isEmpty ()) {
         return;
@@ -67,65 +70,82 @@ void kpToolWidgetBase_Ribbon::saveSelectedAsDefault () const
     cfg.sync ();
 }
 
-void kpToolWidgetBase_Ribbon::relayoutOptions ()
+void kpToolWidgetBase::relayoutOptions ()
 {
 }
 
-int kpToolWidgetBase_Ribbon::selectedRow () const
+int kpToolWidgetBase::selectedRow () const
 {
     return this->selected();
 }
 
-int kpToolWidgetBase_Ribbon::selectedCol () const
+int kpToolWidgetBase::selectedCol () const
 {
     return 0;
 }
 
-int kpToolWidgetBase_Ribbon::selected () const
+int kpToolWidgetBase::selected () const
 {
+    fprintf(stderr, "\tn selected: %d\n", m_galleryGroup->selectionModel()->selectedIndexes().count());
+
     if (m_galleryGroup->selectionModel()->selectedIndexes().count() > 0)
+    {
+        fprintf(stderr, "\t          : %d\n", m_galleryGroup->selectionModel()->selectedIndexes().at(0).row());
         return m_galleryGroup->selectionModel()->selectedIndexes().at(0).row();
+    }
     else
-        return -1;
+        return /*-1*/0;
 }
 
-bool kpToolWidgetBase_Ribbon::hasPreviousOption (int *row, int *col) const
+bool kpToolWidgetBase::hasPreviousOption (int *row, int *col) const
 {
     int sel = this->selected();
 
     if (sel == -1 || sel == 0)
     {
-        *row = *col = -1;
+        if (row)
+            *row = -1;
+        if (col)
+            *col = -1;
+
         return false;
     }
     else
     {
-        *row = sel - 1;
-        *col = 0;
+        if (row)
+            *row = sel - 1;
+        if (col)
+            *col = 0;
         return true;
     }
 }
 
-bool kpToolWidgetBase_Ribbon::hasNextOption (int *row, int *col) const
+bool kpToolWidgetBase::hasNextOption (int *row, int *col) const
 {
     int sel = selected();
     int count = m_galleryGroup->groupModel()->rowCount(QModelIndex());
 
     if (sel == -1 || sel == count - 1)
     {
-        *row = *col = -1;
+        if (row)
+            *row = -1;
+        if (col)
+            *col = -1;
         return false;
     }
     else
     {
-        *row = sel + 1;
-        *col = 0;
+        if (row)
+            *row = sel + 1;
+        if (col)
+            *col = 0;
+
         return true;
     }
 }
 
 // public slot virtual
-bool kpToolWidgetBase_Ribbon::setSelected (int row, int col, bool saveAsDefault)
+bool kpToolWidgetBase::setSelected (int row, int col, bool saveAsDefault)
 {
     setSelected(row, col);
 
@@ -136,7 +156,7 @@ bool kpToolWidgetBase_Ribbon::setSelected (int row, int col, bool saveAsDefault)
 }
 
 // public slot
-bool kpToolWidgetBase_Ribbon::setSelected (int row, int col)
+bool kpToolWidgetBase::setSelected (int row, int col)
 {
     m_galleryGroup->selectionModel()->clearSelection();
     m_galleryGroup->selectionModel()->select(
@@ -150,7 +170,7 @@ bool kpToolWidgetBase_Ribbon::setSelected (int row, int col)
 }
 
 // public slot
-bool kpToolWidgetBase_Ribbon::selectPreviousOption ()
+bool kpToolWidgetBase::selectPreviousOption ()
 {
     int newRow, newCol;
     if (!hasPreviousOption (&newRow, &newCol)) {
@@ -161,7 +181,7 @@ bool kpToolWidgetBase_Ribbon::selectPreviousOption ()
 }
 
 // public slot
-bool kpToolWidgetBase_Ribbon::selectNextOption ()
+bool kpToolWidgetBase::selectNextOption ()
 {
     int newRow, newCol;
     if (!hasNextOption (&newRow, &newCol)) {
@@ -171,26 +191,9 @@ bool kpToolWidgetBase_Ribbon::selectNextOption ()
     return setSelected (newRow, newCol);
 }
 
-void kpToolWidgetBase_Ribbon::callOptionSelected()
+void kpToolWidgetBase::callOptionSelected()
 {
     Q_EMIT optionSelected(selectedRow(), selectedCol());
 }
 
-// protected virtual [base QWidget]
-bool kpToolWidgetBase_Ribbon::event (QEvent *e)
-{
-    return QWidget::event (e);
-}
 
-// protected virtual [base QWidget]
-void kpToolWidgetBase_Ribbon::mousePressEvent (QMouseEvent *e)
-{
-    // delete me
-}
-
-// protected virtual [base QWidget]
-void kpToolWidgetBase_Ribbon::paintEvent (QPaintEvent *e)
-{
-    // Draw frame first.
-    QFrame::paintEvent (e);
-}
